Add files via upload
The AdminControlled Solidity contract uses an onlyAdmin modifier to restrict certain functions to the administrator. The admin variable stores the admin's address. The constructor sets this admin to the address that deployed the contract. The onlyAdmin modifier ensures only the admin can call specific functions. The setNewAdmin function allows the admin to update the admin address, while the getAdmin function lets anyone retrieve the current admin address. This contract demonstrates how to use modifiers for secure access control in Solidity.